requirements.txt
beautifulsoup4~=4.9.3
Pillow~=8.3.1
python-dotenv~=0.18.0
requests~=2.26.0
telethon~=1.23.0
wget~=3.2
youtube_dl
youtube-search-python



